+++ frontend/src/drivers/supaplex/formats/sp/index.ts
@@ -2,6 +2,7 @@ import { ISupaplexFormat } from "../../types";
 import { DAT } from "../dat";
 import { readLevelset, writeLevelset } from "./io";
 import { createLevelset } from "../../levelset";
+import { SIGNATURE_MAX_LENGTH } from "../std";
 import { supportReport } from "./supportReport";
 
 export const SP: ISupaplexFormat = {
@@ -11,6 +12,7 @@ export const SP: ISupaplexFormat = {
   minLevelsCount: 1,
   maxLevelsCount: 1,
   demoSupport: true,
+  signatureMaxLength: SIGNATURE_MAX_LENGTH,
   supportReport,
   readLevelset,
   writeLevelset,